requirements.txt
transformers
sentence-transformers
faiss-cpu
pandas
datasets
torch
